+++ packages/procaptcha-bundle/src/util/widgetFactory.ts
@@ -121,6 +121,7 @@ class WidgetFactory {
 			this.i18n,
 			invisible,
 			widgetContainer,
+			container,
 		);
 
 		return captchaRoot;