+++ packages/types-env/src/env.ts
@@ -18,7 +18,7 @@ import type { KeyringPair } from "@prosopo/types";
 import type { AssetsResolver, EnvironmentTypes } from "@prosopo/types";
 import type { ProsopoBasicConfigOutput } from "@prosopo/types";
 import type { IProviderDatabase } from "@prosopo/types-database";
-import type { IGeolocationService } from "./geolocation.js";
+import type { IIpInfoService } from "./ipinfo.js";
 
 export interface ProsopoEnvironment {
 	config: ProsopoBasicConfigOutput;
@@ -29,7 +29,7 @@ export interface ProsopoEnvironment {
 	keyring: Keyring;
 	pair: KeyringPair | undefined;
 	authAccount: KeyringPair | undefined;
-	geolocationService: IGeolocationService;
+	ipInfoService: IIpInfoService;
 	getDb(): IProviderDatabase;
 	isReady(): Promise<void>;
 	importDatabase(): Promise<void>;