+++ packages/types-env/src/index.ts
@@ -13,4 +13,4 @@
 // limitations under the License.
 export * from "./env.js";
 export * from "./provider.js";
-export * from "./geolocation.js";
+export * from "./ipinfo.js";