+++ dev/flux/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @prosopo/flux
 
+## 2.6.49
+### Patch Changes
+
+  - @prosopo/common@3.1.33
+  - @prosopo/dotenv@3.0.38
+
+## 2.6.48
+### Patch Changes
+
+  - @prosopo/common@3.1.32
+  - @prosopo/dotenv@3.0.37
+
+## 2.6.47
+### Patch Changes
+
+- Updated dependencies [4a9c518]
+  - @prosopo/common@3.1.31
+  - @prosopo/dotenv@3.0.36
+
 ## 2.6.46
 ### Patch Changes